Alert-Create
Alert-CreateVD

+++ AccountingSystem/Models/Product.cs
@@ -3,7 +3,7 @@
     //放公司資料表的資料用
     public class Product
     {
-        public int Product_ID { get; set; }
+        public byte Product_ID { get; set; }
         public string Product_Name { get; set; }
     }
 }
